Servo correctly driven by all peripherals. Button task switching works about 99% of the time. Some intrmittent communication errors or rtos freezes.

+++ CM7/Core/Inc/servoSail.h
@@ -2,12 +2,14 @@
 #define SERVO_SAIL_H
 
 #include <stdint.h>
+#include "FreeRTOS.h"
+#include "task.h"
 #include "cmsis_os2.h"
 
 extern TaskHandle_t task_servoSail;
 
 void servoSail_hardwareInit(void);
 void servoSail_handler(void *argument);
-void servoSail_setAngle(int angle);
+void servoSail_setAngle(int16_t angle);
 
 #endif
